requirements/requirements_dev.txt
django-extensions==1.9.9
django-organizations==1.0.0
et-xmlfile==1.0.1
itsdangerous==0.24
jdcal==1.3
openpyxl==2.5.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
typing==3.6.2
